+++ RELEASE_NOTES.md
@@ -1,2 +1,4 @@
 v0.78.3
 
+- keytrans: Distinguish self-monitoring from monitoring for others
+